refactor(BEB-46): move registro em evento para um modulo separado
