+++ src/modules/events/controllers/index.ts
@@ -3,4 +3,3 @@ export * from './read-events.controller';
 export * from './read-event-by-id.controller';
 export * from './detete-event.controller';
 export * from './read-avaliable-events.controller';
-export * from './event-registration.controller';
