+++ src/modules/events/decorators/index.ts
@@ -4,4 +4,3 @@ export * from './api-read-event-by-id-responses.decorator';
 export * from './api-update-event-responses.decorator';
 export * from './api-delete-event-responses.decorator';
 export * from './api-read-avaliable-events-responses.decorator';
-export * from './api-event-registration.responses.decorator';
