+++ src/modules/events/entities/index.ts
@@ -1,2 +1,1 @@
-export * from './event-atendee.entity';
 export * from './event.entity';
