+++ src/modules/events/input/index.ts
@@ -1,4 +1,3 @@
 export * from './create-event.input';
 export * from './events-filters';
 export * from './update-event.input';
-export * from './event-registration.input';
