+++ src/modules/events/repositories/index.ts
@@ -1,2 +1,1 @@
 export * from './event.repository';
-export * from './events-atendees.repository';
