+++ src/modules/events/services/index.ts
@@ -4,4 +4,3 @@ export * from './read-event-by-id.service';
 export * from './update-event.service';
 export * from './delete-event.service';
 export * from './read-avaliable-events.service';
-export * from './event-registration.service';
