+++ src/modules/registration/controllers/event-registration.controller.ts
@@ -10,9 +10,9 @@ import {
 import { Response } from 'express';
 
 import { UserGuard } from '../../../shared/auth/guard';
+import { EventOutput } from '../../events/output';
 import { ApiEventRegistrationResponses } from '../decorators';
 import { EventRegistrationInput } from '../input';
-import { EventOutput } from '../output';
 import { EventRegistrationService } from '../services';
 
 @Controller('events')
