+++ src/modules/registration/controllers/index.ts
@@ -1,0 +1,1 @@
+export * from './event-registration.controller';
