+++ src/modules/registration/decorators/api-event-registration.responses.decorator.ts
