+++ src/modules/registration/decorators/index.ts
@@ -1,0 +1,1 @@
+export * from './api-event-registration.responses.decorator';
