+++ src/modules/registration/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './event-atendee.entity';
