+++ src/modules/registration/input/event-registration.input.ts
