+++ src/modules/registration/input/index.ts
@@ -1,0 +1,1 @@
+export * from './event-registration.input';
