+++ src/modules/registration/repositories/events-atendees.repository.ts
