+++ src/modules/registration/repositories/index.ts
@@ -1,0 +1,1 @@
+export * from './events-atendees.repository';
